Symlink asm-offsets files to avoid circular dependency.
The asm-offsets header needs to be in an asm/ subdirectory somewhere, so it's easiest to put it here because it's already in the include path. However, it's generated from files under arch/x86/kernel/, and other files in arch/x86/kernel/ need to include this file. So by symlinking the c files we can avoid a circular dependency when parsing Tupfiles.