Lots of Tupfile fixes (link-order, command lines).
These were all found while trying to figure out why initrd wasn't working. In debugging that, I was trying to get my output to be as identical as possible to the make output. Unfortunately the fact that some filenames get embedded in the source makes a bit-for-bit copy difficult, so I diffed things by running nm and such and comparing symbol ordering.