Fix UTS_RELEASE hard-coded version.
In theory this should be generated, but meh. I didn't realize it would cause such trouble by replacing it with the string "gittup". It turns out this can get used by glibc, which expects it to be A.B.C[-something], and in particular the A.B.C need to be valid kernel versions (for example, 2.6.9 and above work, but 2.6.8 causes glibc-built programs to segfault). So my glibc-built "init" test code that I was using in initrd would try to use the bogus version, and then crash. Annoyingly, it wouldn't print anything. After trolling through the entirety of the kernel, I eventually narrowed it down to code in init/, and then finally version.c, and found UTS_RELEASE was the culprit. Apparently the kernel doesn't care what it's set to - it only matters for glibc.